pyproject.toml
[project]
name = "tech-o-health"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "mysql-connector-python>=9.4.0",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "qrcode>=8.2",
    "streamlit>=1.50.0",
    "streamlit-chat>=0.1.1",
]
